requirements.txt
clickhouse-driver
python-dotenv
pandas
tqdm